pyproject.toml
[tool.poetry]
name = "tg-sender"
version = "0.1.0"
description = ""
authors = ["alekseik1 <1alekseik1@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
telethon = "^1.37.0"
python-dotenv = "^1.0.1"
aiogram = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
